Memory optimisation - short index containers

+++ src/main/java/com/dbn/common/index/Indexable.java
@@ -1,5 +1,5 @@
 package com.dbn.common.index;
 
 public interface Indexable {
-    int index();
+    short index();
 }
